Replace insertChannelsMap with channelsMgr (#5453)
