+++ internal/proxynode/channels_mgr_test.go
@@ -1,8 +1,6 @@
 package proxynode
 
 import (
-	"fmt"
-	"math/rand"
 	"testing"
 
 	"github.com/milvus-io/milvus/internal/msgstream"
@@ -22,41 +20,6 @@ func TestNaiveUniqueIntGenerator_get(t *testing.T) {
 		assert.False(t, ok)
 		exists[g] = true
 	}
-}
-
-type mockMaster struct {
-	collectionID2Channels map[UniqueID]map[vChan]pChan
-}
-
-func newMockMaster() *mockMaster {
-	return &mockMaster{
-		collectionID2Channels: make(map[UniqueID]map[vChan]pChan),
-	}
-}
-
-func genUniqueStr() string {
-	l := rand.Uint64()%100 + 1
-	b := make([]byte, l)
-	if _, err := rand.Read(b); err != nil {
-		return ""
-	}
-	return fmt.Sprintf("%X", b)
-}
-
-func (m *mockMaster) GetChannels(collectionID UniqueID) (map[vChan]pChan, error) {
-	channels, ok := m.collectionID2Channels[collectionID]
-	if ok {
-		return channels, nil
-	}
-
-	channels = make(map[vChan]pChan)
-	l := rand.Uint64()%10 + 1
-	for i := 0; uint64(i) < l; i++ {
-		channels[genUniqueStr()] = genUniqueStr()
-	}
-
-	m.collectionID2Channels[collectionID] = channels
-	return channels, nil
 }
 
 func TestChannelsMgrImpl_getChannels(t *testing.T) {
